requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
ultralytics==8.0.145
python-multipart==0.0.6
pillow==10.0.0
numpy==1.25.2
opencv-python-headless==4.8.0.76
opencv-python==4.8.0.76
pydantic==2.3.0
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0 